Post controller Update
Added pagination to [GET] request
Also realized needed methods to sort posts

+++ Practice/Data/Dto/PostCreateDto.cs
@@ -5,6 +5,6 @@
         public string Title { get; set; }
         public string Body { get; set; }
         public List<Guid> TagGuidList { get; set; }
-        public long ReadTime { get; set; }
+        public int ReadTime { get; set; }
     }
 }
