+++ Practice/Data/Dto/PostEditDto.cs
@@ -5,7 +5,7 @@
         public Guid PostId { get; set; }
         public string Title { get; set; }
         public string Text { get; set; }
-        public long ReadTime { get; set; }
+        public int ReadTime { get; set; }
         public List<Guid> AddTagGuidList { get; set; }
         public List<Guid> DeleteTagGuidList { get; set; }
